Fix support for LibreSSL versions other than 2.5.0
